REFACTOR addon-knobs so it will not include runtimes of other platforms
When using knobs on a storybook for react, it should not include anything other then react code.
